+++ OnlineStore.Logic/Queries/Sale/GetUserSale/SaleVm.cs
@@ -3,7 +3,7 @@
     public class SaleVm
     {
         public string UserName {  get; set; }
-        public uint TotalSum {  get; set; }
+        public double TotalSum {  get; set; }
         public DateTime DateTime { get; set; }
 
     }
