+++ OnlineStrore/Dto/UpdateProductDto.cs
@@ -4,7 +4,7 @@
     {
         public string Name { get; set; }
 
-        public uint? Cost { get; set; }
+        public double? Cost { get; set; }
 
         public uint? CountOfProduct { get; set; }
     }
